Move tests to separate folder, slight refactoring of tests and yarn scripts
