+++ .gitignore
@@ -1,2 +1,3 @@
+js
 node_modules
 *.map
