+++ js/internal_modules/identity.js
@@ -3,8 +3,9 @@ Object.defineProperty(exports, "__esModule", { value: true });
 var bs58 = require("bs58");
 var DID = (function () {
     function DID(publicKey) {
-        var fragment = bs58.encode(publicKey);
-        this.identifier = 'did:jolo:' + fragment;
+        var prefix = 'did:jolo:';
+        var suffix = bs58.encode(publicKey);
+        this.identifier = prefix + suffix;
     }
     return DID;
 }());
